Changed quite a bit (READ description)
Deleted uneeded stuff, added (at least attempted to) pause
functionality.  Didn't manage to test it because something went wrong
with the game colors.  You need to look at the colors of everything and
make sure everything is good because when I deleted the random colors
everything screwed up.  Other than that the pause functionality needs
testing.  I'm not sure if I implemented the sending of the isPaused
variable correctly or not so check that too.