Multiple changes (read description)
Fully added the pausing of game, can be done by player one or player two
and it will display on screen who paused the game on both screens. Fixed
the color bug that Tim was having and everything is clearly visible now.
Still some improvements could be made to the pausing as the text isn't
quite properly working yet!