Finished ball class and some smaller changes
I hate git... (But READ THE PREVIOUS COMMIT FOR MOST OF THE
CHANGES!!!!!!)

Fully implemented the ball class with full encapsulation and everything
works.  So all the getters and setters are all working in the main
class.

I spent a while trying to figure out how to properly end a thread since
it's screwing up being able to close the game since the threads wont
stop.  But ending threads (the right way) seems tricky so I think I'll
ask my prof about it tomorrow.  I only tested everything in singleplayer
by the way.