Lots of new stuff again. Still not done :(
Almost fully added GuiText and rendering it correctly, will add buttons
and all that tomorrow if I have enough time as I have some school work
and I have to make my own seperate game for school too. Sigh. Anyway, I
added all of the text but there is a bug with changing the size of the
text, It starts to draw weird lines when resizing, there was someone who
fixed the bug but they didn't have to increase the size of the
textures... so I'm thinking I'll just have to re-draw all of the letters
:/ Good day to you