Half done with the power ups. Various fixes
Fixed a few things like multiplayer working. Added ability to host the
game. When selecting game modes the selected game mode will appear to
stay "selected" added this so the player has a feel of which difficulty
they are going to play on. If the client is a server then by pressing B
increases other players paddle size, this is just a test to see how
powerups will be done. I decided that all the power up bonuses will be
added from the server and not from both clients. Both clients how ever
will still render the power ups, it's just that the math and
calculations on if the ball is touching the power up and who hit the
ball last will be done in the Server class :) Have a nice day and I hope
the code is OK