Small changes
I didn't work on this for very long.  But I made a variable
paddleStopSpeed to make that easier to change, and I spent a while
trying to figure out the bug where the paddle doesn't stop at the
boundries.  I couldn't figure it out but I think it has something to do
with the whole "if (key down){.....}else{ p1dy *= stop}" part.
Hopefully that made sense to you.